% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dftau}
\alias{dftau}
\title{Compute DF Tau Statistic}
\usage{
dftau(dy, x)
}
\arguments{
\item{dy}{Numeric vector of first differences of residuals.}

\item{x}{Matrix of lagged residuals and differenced lags.}
}
\value{
A list containing:
  \item{tau}{The tau statistic (t-statistic for the first coefficient)}
  \item{s2}{Sum of squared residuals}
}
\description{
Computes the Dickey-Fuller tau statistic for unit root testing.
}
\keyword{internal}
