% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks_1_2.R
\name{mbreak1}
\alias{mbreak1}
\title{Detect One Structural Break}
\usage{
mbreak1(datap, n, model, tb, lagoption)
}
\arguments{
\item{datap}{Data matrix with dependent variable in first column.}

\item{n}{Sample size.}

\item{model}{Model specification (0-3).}

\item{tb}{Trimming parameter (number of observations).}

\item{lagoption}{Lag selection option (0 or 1).}
}
\value{
List with tau statistic and break point location.
}
\description{
Internal function to detect one structural break in the cointegration relationship.
}
\keyword{internal}
