% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks_1_2.R
\name{mbreak21}
\alias{mbreak21}
\title{Detect Second Break Before First Break}
\usage{
mbreak21(datap, n, model, tb, bp1, lagoption)
}
\description{
Detect Second Break Before First Break
}
\keyword{internal}
