% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{opttlag}
\alias{opttlag}
\title{Determine Optimal Lag Using t-sig Criterion}
\usage{
opttlag(e)
}
\arguments{
\item{e}{Numeric vector of residuals.}
}
\value{
Integer indicating the optimal lag length.
}
\description{
Determines the optimal lag length using the t-significance criterion
with a maximum lag of 12.
}
\keyword{internal}
